requirements.txt
# FastAPI & ASGI
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Google Gemini
google-generativeai==0.3.0

# AssemblyAI
assemblyai==0.17.0

# Environment variables
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Optional: For production
gunicorn==21.2.0